+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- LocalStorage `orderform` being called with the wrong key
+
 ## [2.67.1] - 2023-05-05
 ### Fixed
 - Fixes of i18n on readme.md accorrding to task LOC-10558.